app/backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-on-heroku==1.1.2
django-restframework==0.0.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
ffmpeg-python==0.2.0
filelock==3.8.0
future==0.18.2
glob2==0.7
huggingface-hub==0.10.1
idna==3.4
more-itertools==9.0.0
numpy==1.23.4
openai==0.24.0
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pandas-stubs==1.5.1.221024
psycopg2-binary==2.9.5
pydotenv==0.0.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
semantic-version==2.10.0
setuptools-rust==1.5.2
six==1.16.0
sqlparse==0.4.3
tokenizers==0.13.1
torch==1.13.0
tqdm==4.64.1
transformers==4.24.0
types-pytz==2022.5.0.0
typing_extensions==4.4.0
urllib3==1.26.12
whitenoise==6.2.0
